+++ src/components/textarea/textarea.stories.tsx
@@ -26,7 +26,7 @@ type Story = StoryObj<typeof Textarea>;
 
 export const DefaultStory: Story = () => {
   const [state, setState] = useState("");
-  const setValue = ({ target }: React.ChangeEvent<HTMLInputElement>) => {
+  const setValue = ({ target }: React.ChangeEvent<HTMLTextAreaElement>) => {
     setState(target.value);
   };
   return <Textarea label="Textarea" value={state} onChange={setValue} />;
@@ -265,16 +265,18 @@ export const BorderRadiusStory: Story = () => {
   const [stateTwo, setStateTwo] = useState("");
   const [stateThree, setStateThree] = useState("");
   const [stateFour, setStateFour] = useState("");
-  const setValueOne = ({ target }: React.ChangeEvent<HTMLInputElement>) => {
+  const setValueOne = ({ target }: React.ChangeEvent<HTMLTextAreaElement>) => {
     setStateOne(target.value);
   };
-  const setValueTwo = ({ target }: React.ChangeEvent<HTMLInputElement>) => {
+  const setValueTwo = ({ target }: React.ChangeEvent<HTMLTextAreaElement>) => {
     setStateTwo(target.value);
   };
-  const setValueThree = ({ target }: React.ChangeEvent<HTMLInputElement>) => {
+  const setValueThree = ({
+    target,
+  }: React.ChangeEvent<HTMLTextAreaElement>) => {
     setStateThree(target.value);
   };
-  const setValueFour = ({ target }: React.ChangeEvent<HTMLInputElement>) => {
+  const setValueFour = ({ target }: React.ChangeEvent<HTMLTextAreaElement>) => {
     setStateFour(target.value);
   };
   return (
@@ -321,7 +323,7 @@ BorderRadiusStory.storyName = "Border Radius";
 
 export const BorderlessExample: Story = () => {
   const [state, setState] = useState("");
-  const setValue = ({ target }: React.ChangeEvent<HTMLInputElement>) => {
+  const setValue = ({ target }: React.ChangeEvent<HTMLTextAreaElement>) => {
     setState(target.value);
   };
   return (
@@ -346,7 +348,7 @@ BorderlessExample.storyName = "Borderless Example";
 
 export const ResizeStory: Story = () => {
   const [state, setState] = useState("");
-  const setValue = ({ target }: React.ChangeEvent<HTMLInputElement>) => {
+  const setValue = ({ target }: React.ChangeEvent<HTMLTextAreaElement>) => {
     setState(target.value);
   };
   return (
@@ -379,7 +381,7 @@ ResizeStory.storyName = "Resize";
 
 export const SizeStory: Story = () => {
   const [state, setState] = useState("");
-  const setValue = ({ target }: React.ChangeEvent<HTMLInputElement>) => {
+  const setValue = ({ target }: React.ChangeEvent<HTMLTextAreaElement>) => {
     setState(target.value);
   };
   return (